# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Takker Swagger Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get products list",
                "description": "Returns a list of products",
                "operationId": "1bfaa78d1c2c3848ab8165c5dadcad3e",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product.",
                "operationId": "storeProduct",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrUpdateProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "String",
                                            "example": "Product created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "integer",
                                            "example": "The tap interval field must be a number."
                                        },
                                        "errors": {
                                            "properties": {
                                                "tap_interval": {
                                                    "type": "array",
                                                    "items": {
                                                        "example": "The tap interval field must be a number"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update existing product",
                "description": "Returns updated product data",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrUpdateProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product successfully updated"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete existing product",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Product": {
                "title": "Product",
                "description": "Product Model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "Product name",
                        "type": "string",
                        "example": "One-Time Payment"
                    },
                    "description": {
                        "title": "description",
                        "description": "Product description",
                        "type": "string",
                        "example": "One full week of premium services."
                    },
                    "price": {
                        "title": "price",
                        "description": "Product price",
                        "type": "number",
                        "format": "float",
                        "example": "30"
                    },
                    "duration": {
                        "title": "duration",
                        "description": "Product duration",
                        "type": "string",
                        "example": "Week"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Product"
                }
            },
...
            "CreateOrUpdateProductRequest": {
                "title": "CreateOrUpdateProductRequest",
                "description": "Product Request",
                "required": [
                    "name",
                    "description",
                    "price"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Product name",
                        "type": "string",
                        "example": "One-Time Payment"
                    },
                    "description": {
                        "title": "description",
                        "description": "Product description",
                        "type": "string",
                        "example": "One full week of premium services."
                    },
                    "price": {
                        "title": "price",
                        "description": "Product price",
                        "type": "number",
                        "format": "float",
                        "example": "30"
                    },
                    "duration": {
                        "title": "duration",
                        "description": "Product duration",
                        "type": "string",
                        "example": "Week"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CreateOrUpdateProductRequest"
                }
            },
...
            "ProductResource": {
                "title": "ProductResource",
                "description": "UserProfile resource",
                "properties": {
                    "message": {
                        "title": "message",
                        "description": "Simple message from api.",
                        "type": "string",
                        "example": "Success operation"
                    },
                    "data": {
                        "title": "data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Product"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ProductResource"
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    }
}